refactor(base.CreateDataset, PrepareDatasetSave): read message for important changes
`base.CreateDataset` accepts a previous dataset and previous body, with lines up with the new `dataset.CreateDataset`, which has new previous dataset and previous body params. Otherwise it is largely unchanged.

`PrepareDatasetSave` has undergone two important changes.

First, removed the commit creation in `PrepareDatasetSave`. Previously, this commit was referenced later down in the stack (in `dataset.prepareDataset`,) to determine what the commit message should be. Now, all of the commit message determination has been isolated to `dataset.generateCommitMsg`, so we no longer need to worry about passing an initial message here.

Second, we now return a second `mutable` dataset. Since we need to pass a clean, unchanged, version of previous to the `dataset.CreateDataset` function, we need a throwaway, mutable version of previous to add to the requested changes in order to come up with the current dataset that will get passed to `dataset.CreateDataset`.  This is the Dataset we will use to populate the "changes" dataset in `actions.SaveDataset`. We need the `prev` dataset to not mutate, because this the datase that we use to compare the current dataset in order to determine what has changed. This can be removed if we ever have a way to deep copy datasets.

Updated tests to match changes!