fix(export): Export --zip flag
Reenable the --zip flag. If the --zip flag is not given, the default format
is "json", but if it is given then the default format is "zip". If --zip is
combined with a format aside from the "zip", then that format is used but the
single-file dataset is inside of a zip file.