refactor(lib.ConfigMethods): remove `ConfigMethods` interface, export `configMethods` struct
We shouldn't have to impliment `ConfigMethods` in a different way, so there is no reason to have `ConfigMethods` be an interface. We should just flip it to being a struct. This also fixes a bug where RPC was not able to connect because the handlers must be on something that is exported.
When we had `ConfigMethods` as an interface, the `configMethods` struct was private.