feat(peers): fix limit/offset bugs, add sending output to `less` when listing peer cache
While working on offset and limit for `qri peers list`, I realized that paginating through the list of peers doesn't really make sense, because to generate the peers list we range through a map, not a slice, we get a different order each time.

We can talk about whether we want to make this list ordered, but for now, I thought it might be a good place to try to stdout to `less`, rather then return the output to the terminal.

So, if you look for your cached peers, qri will send you to `less` that displays your cached peers.