refactor(lib): adjust SaveParams to not require a dataset pointer
lib.SaveDataset used to _require_ a *dataset.Dataset pointer to work, which is an issue for update, which needs to retain a set of input paremeters to "animate" the command line invocation of lib.SaveDatset.

This refactor makes it possible create a dataset using only input paremeters, and moves the initial dataset construction down into lib.SaveDataset. It should now be possible to refactor-away lib.UpdateParams, and store a version of SaveParams to store update input