refactor(update): update service runs via os command invocations
previous iterations got a little ahead of ourselves by attempting to construct an in-process instance to use for dataset update jobs. This refactor introduces a factory pattern to enable that technique, but leaves it disabled for now. Instead cron operates via command line invocations, which brings the memory overhead of the qri update service to 6MB on my machine. This is down from a consistent ~360MB when running the in-process version.