fix: plumb context into network methods
so, turns out cmd package tests that ran after an rpc server starts were all running over rpc, which is bad. This makes that not bad by just fixing the problem, passing in a context that cancels with the test. Its a change we need to make anyway, just not one I thought we'd need to make this go-round. programming is hard.