refactor(repo): remove repo.Path in favour cmd.PathFactory
@dustmop put forth the view that default path negotiation should be the concern of cmd.
Sound right to me, which implies that lib.NewInstance should require a path, and cmd
should do the work of deciding between flags, $QRI_PATH, and the default repo location,
passing that off to lib. Thankfully we already to do this with our cmd.PathFactory
interface, and have a hot-swappable version for testing. This commit removes the
recently-introduced repo.Path method, and does some minor cleanup