refactor(lib): require a context for NewInstance
While we're playing with the signature of NewInstance, it makes sense to also require a context.
library consumers should always be providing a context to the instance constructor, even if just
context.Background(). Adding this requirement signals to the user that they should first
understand how context works before using lib, and that library authors will need to repsect
the passed in context throughout packages below lib.