feat(update): experimental support for multi-repo updates with --use-repo
this is a somewhat advanced use-case, but one that I'd like to address for my
own qri needs (I maintain a number of qri repos on my machine). running
'qri update schedule --use-repo /path/to/other/repo user/dataset R/PT1W'
will store a "repoPath" value along with the job, run qri update with that repo,
which will cause updates to be written and stored in that repo, while running
against the same daemonized update process. Basically: 1 update process,
lots of repos.

Also, introduce a number of fixes caused by a downstream change. Turns out
some of our tests were relying on a broken behaviour: both the manual transform
and ds.set_body had a chance to affect dataset.structure. This is now an error,
and is fixed in tests.