fix(cmd): don't use FgWhite, breaks light-colored termimals
So, I wanted to solve this problem the "right" way, by making sure we're always
writing ANSI codes that we know will print. After some digging, this turns out
to be a bigger problem than I expected, and the easier answer is to avoid using
colors like FgWhite or FgBlack. Thankfully there are only a few places in our
codebase that use these, and one of them was dead code.