fix(setup): load plugins before attempting to setup IPFS
need to load plugins before attempting to configure IPFS, flatfs is
specified as part of the default IPFS configuration, but all flatfs
code is loaded as a plugin.  ¯\_(ツ)_/¯

This works without anything present in the /.ipfs/plugins/ directory b/c
the default plugin set is complied into go-ipfs (and subsequently, the
qri binary) by default

Our tests missed this because the test suite needs to load all plugins to run
tests in the first place. This is going to be particularly tough to test.
We'll need to either:
* figure out a way to cleanly unload all plugins in IPFS
* use an "external" test of some sort

We should do some research into plugin unloading, we could also consider
using a higher level package to test the "final" setup behaviour, which
may not call plugin loading at initialization time.