fix(fsi): Multiple bug fixes, unit tests.
When creating an initial link, do not write the profileID to .qri-ref, only write peername/ds_name. When a directory already exists in a links data structure, check if that directory was deleted or if it still exists. When updating a link, actually write the new ref to .qri-ref. In status, check the error code of ReadDir. Add unit tests to check these new behaviors.