feat(replace-save dataset): support dataset save without patching from prior verion
actions.SaveDataset has been refactored to accept a struct of switches. That struct has a new switch `Replace` that instructs qri to only use input data when creating a dataset commit. This has the effect of "replacing" the dataset instead of applying the input dataset as modifications to the prior version.

We need this for FSI, which operates on a file-system-reflects-current-state paradigm