feat(mock registry): add support for mock registry config
set registry.location to "mock" and qri will create a mock registry, we'll use this in tests elsewhere.

Long term it'd be nice to not have to bundle testing code with our production binary, but the overhead of maintaining a testing-oriented binary is a little too much at the moment.