fix(add): need to re-initialize add when connecting
if we have an IPFS node instance, node.GoOnline has to make a new
instance to connect properly. If remoteClient retains the reference to the
old instance, we run into issues where the online instance can't "see"
the additions. We fix that by re-initializing the client with the new
instance.

Debugging this one was not fun.