refactor(base.DatasetLog): add 700ms timeout to dataset log to prevent dweb hang
this is a horrible hack to handle long-lived requests when connected to IPFS
if we don't have the dataset locally, this process will take longer than 700 mill, because it'll
reach out onto the d.web to attempt to resolve previous hashes. capping the duration
yeilds quick results. The proper way to solve this is to feed a local-only IPFS store to
this entire function