fix(init): add basic `structure.json` on `init`
When we create a new dataset, use the given format to determine what `formatConfig` options we want to expose to the user.

Create the structure.json based on these options

How do we handle `Separator`? It's a rune and marshals to an integer. Should probably be converted to a string?