refactor(actions, base): add `base.RemoveNVersionsFromStore` func
`base.RemoveNVersionsFromStore` removes n versions of a dataset from the store starting with the most recent version. When n == -1, it will remove all versions. It does not remove the dataset reference from the repo.

This replaces `actions.DeleteDataset`, which was not removing all previous versions, only removing the reference and the blocks of that reference.

This also adds test to ensure that the previous datasets are being removed.