feat(remote): embed log hook oplogs in hook context
I see hooks as part of a request/response lifecycle, which makes it appropriate to embed values in a context for later fetching.

We're doing exactly that for oplogs, so hooks can get access to the logs being pushed without affecting the signature of a remote Hook, which is shared by datasets.