fix(remote): remove dataset
First, this commit renames `RemoveDatasetRef` to `RemoveDataset`. It calls on `base.RemoveNVersionsFromStore` to remove the blocks from the filestore, and uses `repo.DeleteRef` to remove the dataset reference from the repo

`RemoveDataset` calls the `datasetRemovePreCheck` function, if it exists, before it removes the dataset and its history.

`RemoveDataset` calls the `datasetRemoved` function, if it exists, after it removes the datasets and its history