fix(remote): disambiguate AuthorID & AuthorPubKey use in logsync
Also! There's a horrible bug this refactoring uncovered in cmd/fetch_test.go: both peers in the fetch are writing to the _same username_, which is bad. Once this refactoring, the test was returning "not found" because we're doing a *get* by alias reference (formerly "test_peer/test_movies"), and post-fetch _two_ peers had the peername "test_peer". (The test used the same "NewTestRepo" func, which hardcoded the peername). There's actually no way to get the ref in this context.  We should be doing checks for ambiguous head references