fix(remove): don't fail if remove encounters an issue traversing qfs history
this is a bit of a stop-gap for now, but users are seeing bad errors when dataset histories aren't fully persisted & they attempt to delete. This may end up causing stray-pinning, and the long-term solution for that is to use logbook for this remove task. We can also use logbook to identify stray pins & reclaim the space

closes #989