feat(lib): add `stats.Cache` to the lib instance as `stats`
added `stats` field of type `stats.Stats` to the `InstanceOptions` and `Instance` struct.

`newStats` func is called in `NewInstance`. We create a new Cache based on the `cfg.Stats.Type` given in config (for now only option is "fs"), with a MaxSize of `cfg.Stats.MaxSize` (by default 25MiB). The cache is saved based on the `repoPath`, or `cfg.Stats.Path` if it's given.