fix(stats): `int` was not capturing all integer options, add `int32` or `int64`
when switching on csv entry reader value types, `int` wasn't capturing all the values that were of some integer type. This was causing any integers in csvs to come back as null.

This fix also checks for `int32` and `int64` in `newAccumulator()` and `numericAcc.Write()`

Also adds tests for csv. Added a test for csv with a full json schema, since it looks like our csv readers don't recognize "null", "NULL", or empty as null in csv.