refactor(stats): allow stats calculation even if there is no path
The `dataset.Path` requirement is only to check the cache to see if we have calculated these stats already. If we have no path, we assume we must re-calculate the stats. Later, we will add specifications for how to cache stats that come from fsi that do not have an associated ipfs path.