refactor(logbook): move author from oplog to logbook
on initial implementation I'd intented for oplog to be a "thicker" API that had a strong notion of authorship. In practice the details of author attribution have happened at the logbook layer as as part of declarative method names. This refactor makes that change explicit, removing private key and author details from oplog.

I've also taken this chance to move the author stuff into the identity package, where it belongs. I'm sure the name "Author" will be refactored over time, but at least it's in the right place.