feat(save): Flag --new for save ensures that we're saving a new dataset
When running `save` without a dataset ref, --new will ensure that a unique
dataset name is generated. When running `save` with a ref, --new will only
succeed if the dataset name is not in use. Add API support so that desktop
can easily import files without having to ask for a dataset name first. Fix
up some inconsistent or incomplete logic in base around saving datasets.