fix(fsi): When working directory is moved or deleted, update FSI Path
Whenever the user manually moves or deletes a working directory, we get into a
situation where the fsiPath in the qri repo is now out of date with the
.qri-ref (or lack thereof) in that working directory. We should detect this
situation as best we can. For now, if the user runs another command in the
folder's new location, use that opportunity to update the fsiPath. Or, if
the user runs `qri list`, check the folders in the repo's references to make
sure everything is still properly linked. In the future, we can watch the
file system to make this process more reliable. But for now, this establishes
the .qri-ref linkfile as the primary source of truth.

Fixes https://github.com/qri-io/qri/issues/938 and https://github.com/qri-io/qri/issues/890