test(cmd): refactor TestRepoRoot down into repo/test package
Now that we have the TestRunner struct in cmd it's possible to refactor a bunch of cmd-specific bits up into that struct, and push the repo mocking bits down so other packages can use it. This kills a long-standing TODO in the process.

I've also merged the lib/test package into the repo/test package. The central magic of a pre-zipped IPFS repo comes from there, and is only being used by cmd tests.