refactor(regserver): update regserver binary to new registry requirements
we're going to need this to properly test our browse experince. And this makes a great chance to work out the API for this stuff.

this commit adds a bunch of capabilities to remotes, which I think is the right call based on our lessons making the initial registry far too thick.