refactor(ref): Move repo.DatasetRef to reporef.DatasetRef in repo/ref
To implement dscache, we need DatasetRef to be plain-old-data with no (or as few) dependencies. Currently it is in the same package as repo, but repo wants to be able to return a dscache. In order to avoid a circular dependency, move DatasetRef to its own package.

As part of this refactoring, split repo/ref.go into repo/ref/dataset_ref.go, repo/parse.go, repo/serialize.go, and repo/store.go.

Some of this is a bit ugly, but it has to happen for dscache, and is only temporary. The next steps:

1) Move the `Dataset` field off of DatasetRef. For call-sites that are currently using it (especially those that require exactly 2 parameters, like rpc calls), make a new struct LoadedDatasetRef that combined a DatasetRef with a Dataset pointer. This way the dependency will at least be explicit.

2) Start using dsref in as many places as possible. I think we can use it for a lot of cmd, with conversions in lib whenever necessary. Switching everything to dsref instead of reporef.DatasetRef is the long-term goal.