refactor(dsref): DetailedRef used in places like List
DetailedRef is a plain-old-data representation of the data returned by methods
like List, which has a dataset ref along with specific details that get
printed to the terminal, or shown in desktop. This removes one major place
that was unnecessarily using reporef.DatasetRef.

This greatly changes the API response, and Desktop will need to adjust how it
is listing datasets. However, it gets Desktop closer to "getting what it
really wants", which will hopefully lead to less code needing to tease apart
API responses.