feat(dscache): Dscache will get updated by save, if it exists in the repo
Add a logbook.Action that describes a change to logbook when a save happens.

Add dscache.Update which receives an action, and modifies the dscache (by copying, since flatbuffers don't support mutation), and saves the result to the repo.

Change dscache constructor so that a non-null dscache is always returned, but it'll just be "empty" if there's nothing in the repo. That allows us to "save" since a filename within the repo should always exist.

Rename --via-dscache flag to --use-dscache.

Make an integration test that demonstrates using `list` to make a dscache, then `save` to update it.