fix(registry): fix inaccurate tests covered by catch-all-200 repsonse regserver handler
commit d3fbce4a0f5590e72364b06c20f55005615f0b03 moved the "healthcheck" handler from "/" to "/health", and in the process uncovered a number of tests that were reporting false-positives. by registring a simple function that responds with a JSON payload of "200 - OK", a number of routes that weren't being created by the registry server were falling back to this OK response.

Because our tests are mainly relying on response codes and not the actual payload of a response, this catch-all reponse made it seem like publication was happening, when in fact nothing was occuring at all. not good. I've addressed the failing tests, but let's make a point of _never_ using the "/" route. On top of that, we should be adjusting our code to plumb server response data back to the user, and checking that data in tests.