fix(remove): RemoveEntireDataset is better at cleaning up broken states
RemoveEntireDataset is a function that cleans up all state related to a dataset in the repository: ipfs pinning, logbook, and the refstore. It continues on despite errors happening, in order to properly clean up when problems happen. Move it to its own file, and add a test that demonstrates how it fixes a problem caused by running `qri init` followed by `rm`ing the working directory. Add better logging to the `Remove` method in order to more easily debug problems like this in the future.