test(cmd): Remove the old cmd integration test
Since it was written, this test has been obsoleted by smaller tests that check the results of individual commands. It isn't particularily covering anything unique. It also has the potential to corrupt a developer's repo due to how it is calling setup. Remove it and continue making integration tests that follow particular user stories.