refactor(fsi): Remove fsi=true parameter, use FSIPath everywhere
Change semantics of lib operations so that, instead of looking for a UseFSI parameter, they see if a reference has an FSIPath. If so, they follow fsi semantics, otherwise refstore. Removes the need to have operations return ErrNoLink if UseFSI is requested but no FSIPath exists.

Most instances of checking for ErrNoHistory now disappear (except for `/history`) since only working directories that have been freshly `init`d will not have history, and now they'll just follow their FSIPath.

It is now illegal to use `status` to check the historical commits of a dataset. Callers should use `whatchanged` instead.

The "fsi=true" api parameter is now non-functional. It now not illegal to send requests with it, it just doesn't do anything.