fix(export): export even if dag contains missing viz referernce
Export has been failing with 'merkledag: not found' thanks to the common culprit of ds.Viz.RenderedPath
not being stored within the underlying dag. To properly fix this, we should instead fix & test the
renderedPath problem, but many datasets have been created with this issue already.

This commit supresses the error generated by referencing a path not found in ds.Viz.RenderedPath,
which has the effect of fixing export.

closes #1161