fix(dsfs): fix dag structure, defend against bad DAGs
we've been seeing a number of "bad dags" that are missing references to hashes
that are intended to be internal to the dag, but are not linked. Use a context
timeout to defend against this.

While we're at it, update the dag structure to write the proper filenames,
which were erroneous. This may not address the central problem

While we're at it, remove "maybeAddDefaultViz" which is a source of problems