refactor(cmd): rename fsi to workdir, add working directory help section
this shuffles things around a bit, but draws a nicer bow around FSI commands
as "working directory commands" in `qri --help`. We've talked about this in
roadmapping documents, and seeing it here, I think it really helps to have
these commands pulled out of the "dataset" list, which is just too long at
this point.

With that said, I don't think `fsi` is a good user-facing term. Working
directory implies what you do: work, in that directory.

closes #1195