fix(setup): Don't crash when running `qri setup`
Initialization and error handling is pretty messy right now. Init can only be run once, but if it encounters an error, that error can very easily be swallowed. Instead, save the err and return it each time Init is called in the future. This allows both accessing the instance early (to assign colors) and also will properly display an error for commands run before setup, instead of having those commands segfault.

In NewQriCommand.PersistentPreRun, check the instance is non-nil before accessing the config.

Remove the call to `f.Init` at the end of Setup, because at that point, Init has already been run and the sync.Once will make calling it again a no-op.