chore: remove update command, subsystem, api
Two problems that lead to the decision to remove `update`. First: it doesn't feel like part of the
core value proposition of qri. There are better scheduler/task managers out there.

Second, it's too buggy, and generally unfinished. We haven't been able to make this work well enough
to justify keeping it, and can't seem to find the bandwidth to justify improving it. We'll solve the
update problem instead with docs on how to wire qri up to a cron job, and may one day revisit
`update` as a plugin/wrapper.

Kill your darlings and whatnot.

BREAKING CHANGE:
update command and all api endpoints are removed