refactor(base): base.SaveDataset takes an initID and path
Require an initID in base.SaveDataset. Now lib.Save has to create an initID first, or look one up for an existing dataset. Doing this requires deciding on a dataset name before entering base, meaning name inference moves up to lib.

CreateDataset does not take an initID, it is only responsible for putting the dataset into block storage and returning the new path. This means it no longer updates logbook. For now, it does update the refstore, so it needs a dataset name, but both of these facts will change in the near future.

Add todos about upcoming changes to transform and apply and dry-run.

Add a testRunner for base.