refactor(diff): Refactor diff for clarity. Use lib.TestRunner
Diff needed a bit of work, as its implementation has been a bit tricky to
understand, and there were some known issues reported recently. We also
want to move some tests down from `cmd` to `lib` and this is a good
opportunity to do so. In addition, make the lib.Diff interface easier to
understand by making explicit the cases that it will handle.