feat(dscache): Dscache used for fsi init and checkout
Dscache is updated by both logbook and fsi using a hook, which is defined
in dsref/hook. This hook represents any sort of dataset change. In order to
make this work properly, dscache should not live on the repo anymore, but
instead should only be on lib.Instance. This is because the lib.Instance
has an FSI struct, but the repo does not (it only has logbook). There are
some hacks in place to workaround this until it is fixed.