refactor(fsi): remove CanonicalizeDatasetRef from FSI
Lots of groundwork for moving past the old stack of DatasetRef, Canonicalize,
and RefStores within the repo interface.

I've also changed the behaviour of fsi.CreateLink to _require_ a dataset
exists before a link can be created. Now that we have logbook identifiers,
to model a state of no history, this is much easier to envision.