feat(`lib`, `cmd`): add wait group and done channel to `lib.Instance` and `cmd.QriOptions`
Following the pattern established in `qfs`, added a wait group and done channel to `lib.Instance` and `cmd.QriOptions`

Since we currently only have one thing we are `waiting` on in both cases, this might seem a little excessive right now. But, since this is a bit of a confusing problem, I'm instead focusing on having it be very easy to add to. Use `waitForOneDone(done chan struct{})` to add another channel to the wait group, and make sure `waitForAllDone()` is running in a go routine somewhere to catch it.

refactor(`repo/test`, `p2p/test`, `base/dsfs`): adjust to new `qfs` api

Some tests expect (specifically tests in `registry/regclient` expect a working ipfs node. No other tests expect it, so those are using cafs map stores for their store and repo.
New `TempRepo.IPFSRepo()` assumes the repo was created with a real qri `Config` and points to an actual ipfs repo. It returns a repo that is ipfs enabled.

- context to `qipfs.NewFS()`, as per new `qfs` api in testing functions in `repo/test`, `p2p/test`, and `base/dsfs/package_test.go`
- add context to `p2ptest.MakeRepoFromIPFSNode` func (passes context down to `qipfs.NewFS`)
- point to latest `qfs` api