refactor(`repo/test/temp_repo`): add `IPFSRepo` option to return a working ipfs repo
Some tests expect (specifically tests in `registry/regclient` expect a working ipfs node. No other tests expect it, so those are using cafs map stores for their store and repo.

New `TempRepo.IPFSRepo()` assumes the repo was created with a real qri `Config` and points to an actual ipfs repo. It returns a repo that is ipfs enabled.