feat(`lib`, `cmd`): add wait group and done channel to `lib.Instance` and `cmd.QriOptions`
Following the pattern established in `qfs`, added a wait group and done channel to `lib.Instance` and `cmd.QriOptions`

Since we currently only have one thing we are `waiting` on in both cases, this might seem a little excessive right now. But, since this is a bit of a confusing problem, I'm instead focusing on having it be very easy to add to. Use `waitForOneDone(done chan struct{})` to add another channel to the wait group, and make sure `waitForAllDone()` is running in a go routine somewhere to catch it.