refactor(`cmd.NewQriCommand`): returns a `done` channel
`cmd.NewQriCommand` returns a `done` channel that you can wait on to be sure that the given qri command has clean up its resourses.

We may have a cleaner way to handle this soon, but currently, every place we create a command runner is bookended by the creation of a a context with a cancel that is passed into the allocation of the command runner.
The command runner returns a channel on which we can wait to ensure we wait until the command is cleaned up before moving on.
We need to manually call `cancel` to ensure the command tears down.