feat(logbook): write push & pull operations, logbook diagnostic method
Prior to this commit we were _not_ recording push & delete operations in logbook
As per RFC0030 https://github.com/qri-io/rfcs/blob/master/text/0030-replace_publish_clone_with_push_pull.md
we're now recording pushes of specific versions to a given remote within the
oplog.

To do this, we make an optimistic write _before_ pushing or deleteing logs that
adds an operation assuming it's successful. This way the receiver of the log
gets the most up-to-date log that is signed with the operation included. If
the push or delete operation fails, we roll back the log, removing the operation.

This approach tolerates faults better. the result of having an extra
push / delete operation is less problematic than having a log that is short by
one operation on a remote.

With this in place, logbook can now accurately report which versions of a
dataset have been published, and has a running record of _where_ those versions
have been published.

I've also added a "Diagnostic" method that's turned out to be very useful for
quickly understanding the contents of a logbook by printing
"MODEL ID OPCOUNT NAME" in an indented hierarchy. This was valuable enough to
merit adding to the logbook command as a flag. Sample output:

```
user icm3gkoaoqn4vpcgzx7rmqsupcdcuec7fgqdq5z4kinvjprshqqq 2 nyc-transit-data
  dataset s3cql3s7fbay7abohm3qqdas7cm4ymu47jplqog5hxos5ytgaywa 2 turnstile_daily_counts_2020
    branch 2bz6mglcqlxa72nb3b7hnz4akrn6swi6iaredvzbgiwf5lp7fg5q 12 main
  dataset w5nl6h7fgwvu6gtfuxjo7fx4lwnerl7uuqycmu2ydrvdx4n76c7a 3 turnstile_daily_counts_2019
    branch 3ave4lm42mym2iuzyazv5b3ackdpvmtredcy4gdgbdsf3ulqxxoq 4 main
  dataset zngx523lbvukv4vebomvpzwsccvab5nuxf5c6h3svjdakpblotea 5 turnstiles_station_list
    branch vyznxsyekkjcfwfpor2svc7v5if2ec7bfq2p5u65xr3xkvnuhqoq 3 main
```