feat(lib): add OptEventHandler for subscribing to events move WS to instance
shift toward subscribing to events at the point of instance construction with
an option to provide an event handler & event type subscription list as a
NewInstance option.

While we're at it, let's internalize the websocket connection, which might
actually be able to replace the RPC api in the long run? either way it
lets a bunch of fields remain unexported.