fix(cmd): re-work migration execution
qri desktop needs more handles for probing when a migration needs to be run.
I've re-worked the logic to make a few new flag & execution combinations
possible, and adjusted exit codes so desktop has an easy hueristic for detecting
migrations from CLI commands. CLI now exits with one of three status cudes:

* 0 - success
* 1 - generic error
* 2 - migration is required

Untangle the `--no-prompt` flag from the `--migration` flag. The intent of
`--no-prompt` is "don't stop the flow of the program for user input, just
keep going". This includes asking users to re-run a command. With no-prompt,
successful migrations should continue.

`--migrate` is now a global flag that means "auto run migrations"

$ qri config get

$ qri config get --no-prompt

$ qri config get --migrate

$ qri config get --migrate --no-prompt

This seems like a lot, but I think it's worth it to have this level of
granularity. Desktop needs the `--no-prompt` command to check if a migration is
required, and the `--migrate --no-prompt` combo to actually run the migrations
and continue.