refactor(repo): repos provide a bus accessor
to properly cleanup "newTempRepo" we need to make sure it has a real event bus.
We're trying to avoid adding things to the repo interface, but there's no way
around this one without moving up into instance land, which isn't possible for
numerous test suites at the moment.

This change uncovered a bug where TempRepo struct was retaining the internal repo,
which would have crossed wires with the provided context argument. This might help
with some of the "repo already open" errors we're seeing in the cmd package tests
(I don't expect this is the *only* source of this error, though).

also: remove NewIPFSRepoFromDir. it has zero callers