refactor(clone): rename clone to pull, use dsrefs in remote, peerIDs in ResolveRef
Too many changes in one commit, but we've got lots going on these days, so such
is life. Most of these changes stem from "qri add" suddenly ceasing to work,
which upon investigation seems like it could stem from a number of problems,
at least some of which derive from blending calls to CanonicalizeDatasetRef and
ResolveRef in the same code path.

It's a bad thing that no test caught this. This commit modernizes the remote
package, setting it up for a more robust package-level test suite. All calls to
CanonicalizeDatasetRef are dropped in favour of ResolveRefs.

Updating the remote package clarifies that ResolveRef really does need to set
the peerID as part of reference resolution. Long term this will need to evolve
as collaboration necessitates the transition to dataset authors using a shared
secret keypair instead of piggybacking on the author's keypair, but either way:
reference resolution needs to set a public key verification mechanism so logbook
data can be confirmed over the wire. This commit makes a spec change enforcing
such a requirement

BREAKING CHANGE:
* all remote hooks pass dsref.Ref instead of reporef.DatasetRef arguments
* ResolveRef spec now requires public key ID be set as part of reference resolution