refactor: changes to make testground simulation viable (#1274)
* refactor(lib): move parts of cmd setup into lib

we're using qri directly as a library in more & more places, and there's
some logic up in cmd that makes setup easier. This change brings automatic
profile generation into lib.

long term, I'm now thinking it'd be smarter to have a default crypto
source available in lib that setup can use, and attach that crypto source
to an instance by default, allowing for an override via instance config.

* chore(p2p): cleanup, remove stale comments in p2p package

* feat(p2p): add event support for p2p, publish on qri peer connect

the p2p subsystem is one of the primary reasons we added support for events in the first
place. Emitting an even on successful peer connection opens up number of "oh hey, it's them! do this!"
type of reactionary code. I've also stubbed in some basic events for going online/offline, and qri peer
disconnect.

Added a few tests to prove events are being emitted at the p2p & lib level

* fix(p2p): don't fail if discovery can't start

found while setting up testground, which doesn't have mdns out of the box

* fix(p2p): use network notifee to be notified of connections when no mdns is present

* refactor(p2p): numerous p2p fixes

Work on testground has unveiled a number of problems that need addressing in our
p2p library. The biggest problem is *still* with UpgradeToQriConnection, which
is firing at the wrong time in a connection lifecycle.

We want to check all new connections for qri support, and upgrade them if we
detect a p2p peer supports Qri. There is a better way to do this, but it's in a
future version of libp2p we don't depend on yet. We're in the process of
upgrading both libp2p *and* IPFS, and should wait for those to land before
moving over to additional refactoring. For now the p2p tests are going to be
*extremely* flaky.

But this commit brings a bunch of cleanup that I think is worth it. namely:

* switch to /p2p/ peer addresses, matching shift coming from IPFS
* remove erroneous context field from QriNode, node.GoOnline now reqires a
  context instead, which is where we should be using contexts
* remove msgChan field from QriNode. It was only being used to write to the list
  of receivers, which is now done directly when we get a message
* only create a discovery service if we're using our own libp2p host (not IPFS)
* adjust discovery handler to only attempt to connect on discovery, matching
  IPFS behaviour. Qri protocol upgrade is handled by network notifee handler
* move "UpgradeToQriConnection" into a network notifee handler. Now all new
  connections will get a uniform check for Qri protocol upgrade
* drop commented-out test: TestRequestProfileOneWayConnection
* add events for peer connect / disconnect

* fix(p2p): more p2p refactoring & fixes

* refactor(p2p/node): upgrade to qri conn at p2p connected event

* fix(instance): re-initialize remote after going online

We reinitialize the remote client with the new ipfs node information after we have gone online, we should also reinitialize the remote with the new ipfs node

* fix(lib.GoOnline): start dsync services when we bring the node online.

* refactor(`ConnectQriNodes`): removing func that no longer does what it promises

`ConnectQriNodes` used to explicitly call `upgradeToQriConnection` to get test nodes to upgrade to a qri connection. Now, however, we expect the nodes to upgrade themselves when the `libp2pevent.EvtPeerIdentificationCompleted` event gets published.

* refactor(lib): add `OptEnableRemote`

* refactor(remote): abstract dysnc setup in remote, use `GoOnline`
This commit also adds back the libp2pevent.BufSize opt

Co-authored-by: ramfox <kasey@qri.io>